cronのエンドポイントを分ける
